Renamed to avoid include conflict in parser
