+++ src/utility2/EventProfiler.cpp
@@ -1,4 +1,4 @@
-#include "utility/EventProfiler.hpp"
+#include "utility2/EventProfiler.hpp"
 
 #include <string>
