+++ src/utility2/EventProfiler.hpp
@@ -17,7 +17,7 @@
 #include <vector>
 
 #include "threading/Mutex.hpp"
-#include "utility/Macros.hpp"
+#include "utility2/Macros.hpp"
 
 #include "glog/logging.h"
