+++ src/utility2/SyncStream.hpp
@@ -5,7 +5,7 @@
 #include <ostream>
 #include <unordered_map>
 
-#include "utility/Macros.hpp"
+#include "utility2/Macros.hpp"
 
 class SyncStream {
  public:
