+++ src/utility2/ThreadSafeQueue.hpp
@@ -8,7 +8,7 @@
 #include <queue>
 #include <utility>
 
-#include "../utility/Macros.hpp"
+#include "../utility2/Macros.hpp"
 
 template <typename T>
 class ThreadSafeQueue {
